Support forcing type on Helm chart `set` attribute
Certain values in a Kubernetes resource _must_ be a string and
Kubernetes will complain if the value is of another type rather than
automatically convert it into a string. An example of this problem would
be the value `true` on a `Pod` label - in YAML, this must be `"true"`
for Kubernetes to be satisfied.

This adds a new optional attribute for the `set` element of all
`*_helm_config` variables which allows the user to switch from the
default behaviour of handling the `value` - `auto` - to `string`, which
will force the `value` to be handled as text rather than a boolean or
number.

Fixes #942