Changing docker files to be stored on secondary drive
The Windows self-managed node is being deployed with an additional 100GB Amazon EBS disk, where the actual Userdata described in this script initializes the disk and formats but doesn't give any usage for the disk. So, to enhance docker performance during node startup by avoiding concurrency disk I/O with the Operational system and image pulling, I'm proposing these additional lines in the scripts that set docker data to be stored on the secondary disk.