Added Windows kubelet best practice
Kubelet behaves differently on Windows, and some parameters need to be in place in order to provide kubelet with an optimal environment.

--System-reserved memory=1.5Gb 
Windows do not have an out-of-memory process killer as Linux does. Windows always treat all user-mode memory allocations as virtual, and pagefiles are mandatory. Windows nodes do not overcommit memory for processes. The net effect is that Windows won't reach out of memory conditions the same way Linux does and processes page to disk instead of being subject to out-of-memory (OOM) termination. If memory is over-provisioned and all physical memory is exhausted, then paging can slow down performance.

--windows-priorityclass=ABOVE_NORMAL_PRIORITY_CLASS
Windows can limit the amount of CPU time allocated for different processes but cannot guarantee a minimum amount of CPU time. On Windows, the kubelet supports a command-line flag to set the scheduling priority of the kubelet process: --windows-priorityclass. This flag allows the kubelet process to get more CPU time slices when compared to other processes running on the Windows host. More information on the allowable values and their meaning is available at Windows Priority Classes. To ensure that running Pods do not starve the kubelet of CPU cycles, set this flag to ABOVE_NORMAL_PRIORITY_CLASS or above.